TabBar Resources, changes to MessageViewController
Added a few icons to make the tabbar not look so empty. 
Made changes to MessageViewController:
Changed textfield to a textview which will dynamically increase with content to take advantage of the multiple line capability of the tableview.
Added observers to monitor for keyboard events. 
Changed query to go for "content" field.